# docs/swagger.yaml
openapi: "3.0.0"
info: 
  title: learn express and swagger
  description: a course segment about writing docs and
  version: 1.1.0
  contact: 
    email: "email@example.com"
    url: "http://example.com"

# servers:
#   - url: "https//localhost:4000/api/v1"
#     description: for localhost - secure
#   - url: "http://localhost:4000/api/v1"
#     description: for localhost - regular

servers:
  - url: "{protocol}://localhost:4000/api/{version}"
    description: for localhost
    variables: 
      version: 
        enum: 
          - v1
          - v2
        default: v1
      protocol: 
        enum: 
          - http
          - https
        default: http

components: 
  securitySchemes: 
    cookieAuth:
      type: apiKey
      in: cookie
      name: token        
    BearerAuth:
      type: http
      scheme: bearer
paths: 
  /course:
    get:
      tags:
        - String
      summary: returns a greet message from course
      responses:
        200:
          description: All good success
          content: 
            application/json:
              schema:
                type: string
                example: "example-string"
        400:
          description: Bad request
        500:
          description: Internal server error
  /courseobject: 
      get:
        tags:
          - Object
        summary: returns a unique course 
        responses:
          200:
            description: All good success
            content: 
              application/json:
                schema:
                  type: object
                  properties: 
                    id:
                      type: string
                    name:
                      type: string
                    price:
                      type: string

          400:
            description: Bad request
          500:
            description: Internal server error
  /courses: 
      get:
        tags:
          - Array
        summary: returns all courses
        responses:
          200:
            description: All good success
            content: 
              application/json:
                schema:
                  type: array
                  items:
                    type: object
                    properties: 
                      id:
                        type: string
                      name:
                        type: string
                      price:
                        type: number

          400:
            description: Bad request
          500:
            description: Internal server error
  /addCourse: 
      post:
        tags:
          - String
        summary: add a new course to existing courses
        consumes:
          - application/json
        produces:
          - application/json
        requestBody:
          required: true
          content: 
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties: 
                    id:
                      type: string
                    name:
                      type: string
                    price:
                      type: number
        responses:
          200:
            description: All good success
            content: 
              application/json:
                schema:
                  type: array
                  items:
                    type: object
                    properties: 
                      id:
                        type: string
                      name:
                        type: string
                      price:
                        type: number
          400:
            description: Bad request
          500:
            description: Internal server error
  /coursequery: 
      get:
        tags:
          - String
        summary: trying to learn about query
        parameters:
          - name: location
            in: query
            required: true
            schema:
              type: string
              enum: [San Francisco, New York, Seattle]
          - name: device
            in: query
            required: true
            schema:
              type: string
              enum: [web, mobile]
        responses:
          200:
            description: All good success
            content: 
              application/json:
                schema:
                  type: object
                  properties:
                    location: 
                      type: string
                    device: 
                      type: string
          400:
            description: Bad request
          500:
            description: Internal server error
  /imageupload:
      post:
       tags:
        - String
       summary: uploading course image
       parameters:
        - in: header
          name: auth
       requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
       responses:
        200:
          description: All good success
              
        400:
          description: Bad request
        500:
          description: internal server error

  